### 2024-09-26 #255 @sambit-giri
@labarba I was wondering if you could kindly provide some insight into how long it typically takes for reviewers to be assigned to a submission.
