+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={1},
   number={1},
   pages={7--22},
+  doi = {10.1108/20468251211179678},
   year={2011},
   publisher={Emerald Group Publishing Limited}
 }
@@ -29,6 +30,7 @@
   booktitle={2012 conference on intelligent data understanding},
   pages={47--54},
   year={2012},
+  doi = {10.1109/CIDU.2012.6382200},
   organization={IEEE}
 }
 
@@ -39,6 +41,7 @@
   volume={938},
   number={2},
   pages={110},
+  doi = {10.3847/1538-4357/ac8e04},
   year={2022},
   publisher={IOP Publishing}
 }
@@ -50,17 +53,22 @@
   volume={152},
   number={6},
   pages={157},
+  doi = {10.3847/0004-6256/152/6/157},
   year={2016},
   publisher={IOP Publishing}
 }
 
-@article{robert2004metropolis,
-  title={The metropolis—hastings algorithm},
-  author={Robert, Christian P and Casella, George and Robert, Christian P and Casella, George},
-  journal={Monte Carlo statistical methods},
-  pages={267--320},
-  year={2004},
-  publisher={Springer}
+@Inbook{robert2004metropolis,
+  author="Robert, Christian P., and Casella, George",
+  title="The Metropolis---Hastings Algorithm",
+  bookTitle="Monte Carlo Statistical Methods",
+  year="2004",
+  publisher="Springer New York",
+  address="New York, NY",
+  pages="267--320",
+  isbn="978-1-4757-4145-2",
+  doi="10.1007/978-1-4757-4145-2_7",
+  url="https://doi.org/10.1007/978-1-4757-4145-2_7"
 }
 
 @phdthesis{giri2019tomographic,
@@ -76,6 +84,7 @@
     author = {Furlanetto, S.~R. and Oh, S.~P. and Briggs, F.~H.},
     month = {10},
     pages = {181--301},
+    doi = {10.1016/j.physrep.2006.08.002},
     volume = {433},
     doi = {10.1016/j.physrep.2006.08.002}
 }
@@ -185,7 +194,7 @@
     month = {10},
     pages = {1590--1605},
     volume = {489},
-    publisher = {Narnia},
+    publisher = {Oxford University Press},
     url = {},
     doi = {10.1093/mnras/stz2224},
     issn = {0035-8711}
@@ -243,6 +252,7 @@
     author = {DeBoer, David R and Parsons, Aaron R and Aguirre, James E and Alexander, Paul and Ali, Zaki S and Beardsley, Adam P and Bernardi, Gianni and Bowman, Judd D and Bradley, Richard F and Carilli, Chris L and {others}},
     number = {974},
     pages = {45001},
+    doi = {10.1088/1538-3873/129/974/045001},
     volume = {129},
     publisher = {IOP Publishing}
 }
