### 2024-09-27 #255 @arm61
Looks good now. A heads up that the LaTeX has mangled the umlaut containing words in your affiliation. Perhaps take a look at that.
