### 2024-09-27 #255 @sambit-giri
> Looks good now. A heads up that the LaTeX has mangled the umlaut containing words in your affiliation. Perhaps take a look at that.

@arm61 Thank you for noticing this. It should be fixed now.
