### 2024-10-08 #261 @editorialbot
[REVIEW]: AstronomyCalc: A python toolkit for teaching Astronomical Calculations and Data Analysis methods
**Submitting author:** @sambit-giri (<a href="http://orcid.org/0000-0002-2560-536X">Sambit Kumar Giri</a>)
**Repository:** https://github.com/sambit-giri/AstronomyCalc
**Branch with paper.md** (empty if default branch):
**Version:** 1.1
**Paper kind:** software
