### 2024-10-08 #261 @kelle
- [x] **Version:** Does the release version given match the GitHub release?
