### 2024-10-30 sambit-giri/AstronomyCalc#1 @kelle
Use astropy.cosmology
Much of the functionality presented here is available in the [Astropy Cosmology sub-package](https://docs.astropy.org/en/stable/cosmology/index.html), which is very well tested and maintained.  This package will be more reliable and sustainable if it relied on `astropy.cosmology` rather than newly built functions.

Submitted as part of JOSE Review.
https://github.com/openjournals/jose-reviews/issues/261
