### 2024-10-30 sambit-giri/AstronomyCalc#1 @kelle
See, for example, this tutorial: https://learn.astropy.org/tutorials/redshift-plot.html
