### 2024-10-30 sambit-giri/AstronomyCalc#2 @kelle
Example data is vulnerable to link rot
The datasets used by this package are quite useful but

1. are hidden to the user by the `create_data` module and
2. are dependent on websites outside of the author's control.

I would recommend removing those datasets and functions as part of the package and instead make them part of the tutorials:
-  https://github.com/behrouzz/astrodatascience/raw/main/data/hubble1929.csv
-  http://astroweb.cwru.edu/SPARC/Rotmod_LTG.zip

If data munging isn't a learning goal, you could also consider re-bundling the data into simpler formats without the excess columns and include those files as part of this repository.

Submitted as part of JOSE Review.
https://github.com/openjournals/jose-reviews/issues/261
