### 2024-10-30 sambit-giri/AstronomyCalc#3 @kelle
Remove constants.py
Astropy should be used for all unit conversions and constants:
- https://docs.astropy.org/en/latest/units/index.html
- https://docs.astropy.org/en/stable/constants/index.html

Submitted as part of JOSE Review.
openjournals/jose-reviews#261
