### 2024-10-30 sambit-giri/AstronomyCalc#4 @kelle
Consider using the `emcee` package
Many of the functions in `mcmc.py` could be replaced by using the widely-used `emcee` package.
https://emcee.readthedocs.io/en/stable/

Submitted as part of JOSE Review.
openjournals/jose-reviews#261
