### 2024-10-30 sambit-giri/AstronomyCalc#5 @kelle
Consider unpackage-ing
The code here is useful for exploring various cosmological concepts however, much of the code in the actual AstronomyCalc package should be replaced with functions available in other widely-used packages such as Astropy and emcee.  Instead of bundling this work as a new package, I would recommend publishing as a collection of notebooks.

Submitted as part of JOSE Review.
openjournals/jose-reviews#261
