### 2024-10-30 sambit-giri/AstronomyCalc#6 @kelle
Contribute to Astropy Tutorials
Once the tutorials are updated to use Astropy, they would be a great addition to the Astropy tutorials!
https://learn.astropy.org/contributing/how-to-contribute

The Astropy Tutorials project has been in a bit of a lull the past couple of years but it has recently gotten a new lead and a new maintainer so we're optimistic to be publishing new notebooks very soon and these would be perfect!

Submitted as part of JOSE Review.
openjournals/jose-reviews#261
