### 2024-12-06 sambit-giri/AstronomyCalc#5 @sambit-giri
I agree that there are several highly optimised packages out there for the calculations presented in the tutorials, which could be good for the students to learn to use. However, the key goal of this package is to show the students how the concepts or equations taught in the classroom can be coded up and give them tasks where they would be asked to use some functions or go beyond it. Using the widely-used packages during teaching may help them learn using these packages but not developing them.
