### 2024-12-06 sambit-giri/AstronomyCalc#5 @sambit-giri
I am considering to instead extend the notebook with advanced examples using the widely-used packages. This would direct them to try out these packages after learning the basic methods. @kelle what do you think?
