### 2024-12-09 #261 @sambit-giri
Hi @kelle,

I have started working on the issues you created.

Most of the issues suggest using another widely-used package for each tutorial notebook instead. While this could help students learn to use these packages, it may not fully achieve our goal of teaching them how to develop and implement these methods from scratch. The primary goal of this package is to demonstrate how the concepts and equations taught in the classroom can be implemented in code. The simple implementations would serve as the starting point that students further develop during the tutorial sessions, with group activities built around this. This approach encourages students to understand the underlying principles and develop their own functions, rather than simply using pre-existing packages.

To balance this, I am considering extending the notebook with advanced examples that utilize widely-used packages. This way, students can first learn the basic methods by coding them from scratch and then explore how these methods are implemented in professional tools. This approach will help them understand both the fundamental concepts and the practical applications. What do you think about this plan?
