### 2024-12-09 2faefed
Modules now depend on astropy units and constants
Fix #3
