+++ setup.py
@@ -12,7 +12,7 @@ with open('requirements.txt') as f:
     requirements = f.read().splitlines()
 
 setup(name='AstronomyCalc',
-      version='1.0.0',
+      version='1.0.1',
       author='Sambit Giri',
       author_email='sambit.giri@gmail.com',
       packages=find_packages("src"),
