+++ src/AstronomyCalc/cosmo_calculator.py
@@ -3,7 +3,6 @@ from scipy.integrate import quad
 
 from .basic_functions import *
 from .cosmo_equations import * 
-from . import constants as const
 
 def age_estimator(param, z):
     '''
@@ -19,8 +18,8 @@ def age_estimator(param, z):
     Feq = FriedmannEquation(param)
     a = z_to_a(z)
     I = lambda a: 1/a/Feq.H(a=a)
-    t = lambda a: quad(I, 0, a)[0]*const.Mpc_to_km/const.Gyr_to_s
-    return np.vectorize(t)(a) 
+    t = lambda a: quad(lambda x: I(x).value, 0, a)[0]
+    return (np.vectorize(t)(a)*I(1).unit).to('Gyr') 
 
 def Hubble(param, z=None, a=None):
     """
