+++ src/AstronomyCalc/constants.py
@@ -1,4 +1,0 @@
-
-c_kmps = 299792.
-Mpc_to_km = 3.0860e19
-Gyr_to_s  = 3.1536e16
