### 2024-12-09 #261 @kelle
I am still a bit confused about why a package is needed if the goal is to teach them "from scratch", but I see the gist of your point. I think your idea of expanding the notebooks to demonstrate the astropy functionality is an excellent one.
