### 2024-12-09 #261 @sambit-giri
The initial modules were created while teaching a bachelor's course introducing students to cosmology. I have found that several similar introductory courses in astronomy do not have basic coding knowledge as a mandatory prerequisite. Therefore, many students in my cosmology course needed some examples to get started. These examples helped them quickly build on their knowledge to solve the assignment questions.

I have tried pointing the students to popular packages, but only some of them could figure it out. I will expand the notebooks introducing other relevant packages (e.g. astropy and emcee) so that the students have the option to learn the advanced tools.
