### 2024-12-16 sambit-giri/AstronomyCalc#5 @sambit-giri
I want to refer to the reply [here](https://github.com/openjournals/jose-reviews/issues/261#issuecomment-2527790360) and the discussion post the reply. Instead of complete unpackage-ing, this package is a collection of several simple implemenation of the methods used in astronomy that can be introduced to students. In the tutorial notebooks, I will mention the more oprimised and widely used packages that the interested students can explore.
