### 2024-12-16 0e1a683
vectorisation problem with astropy units
