### 2024-12-16 sambit-giri/AstronomyCalc#1 @sambit-giri
Thanks for these references. Though the cosmological calculations are similar to the `astropy.cosmology` package, this package is more focused on allowing students to get introduced to how the calculations are perform through the simple implementations. `astropy.cosmology` is better for research work where accuracy is very important.
