### 2024-12-16 300794b
extended to refer to astropy.cosmology for advanced calculations, closes #1
