### 2024-12-16 sambit-giri/AstronomyCalc#4 @sambit-giri
The `mcmc.py` also contains a simple inference method, which is `Importance Sampling`, that is not suitable for real astronomical data but helps give the students an intuition of the posterior distribution.
