### 2024-12-16 sambit-giri/AstronomyCalc#4 @sambit-giri
`emcee` is a very good package for MCMC but not suitable for teaching students what is happening under the hood. Of course, it is very good for them to be aware of such package that they can use later for better data analysis. I have extended the corresponding notebook mentioning several packages, such as `emcee`, `PolyChord` and `sbi`.
