### 2024-12-16 b4ca4e8
plotting pie chart of content of universe
