+++ src/AstronomyCalc/__init__.py
@@ -6,6 +6,7 @@ from .cosmo_calculator import *
 from .cosmo_equations import * 
 from .create_data import *
 from .mcmc import *
+from .plotting_tools import *
 
 # # Re-export modules to make them accessible when the package is imported
 # __all__ = [
