### 2024-12-16 77d1f1f
dataset are shared as a part of the package, fixes #2
