### 2024-12-16 sambit-giri/AstronomyCalc#2 @sambit-giri
Thank you for this suggestion. I have now inlcuded these dataset in the package and will use the same strategy in the future.
