### 2024-12-17 sambit-giri/AstronomyCalc#6 @sambit-giri
Thank you for your thoughtful suggestion! I completely agree that integrating updated tutorials using Astropy into the Astropy Tutorials project would be a fantastic long-term contribution. I appreciate the information about the recent updates to the project and the link to the contribution guide.
