### 2024-12-17 sambit-giri/AstronomyCalc#6 @sambit-giri
I will keep this strategy in mind to ensure my tutorials can become part of the Astropy Tutorials in the future. For now, since this suggestion does not directly impact the current structure or functionality of the package under review, I will close this issue.
