### 2024-12-17 16bba8f
content diagram added
