### 2024-12-24 d8acedc
install notes
