+++ README.md
@@ -48,7 +48,10 @@ One can also install the latest version using pip by running the following comma
 
     pip install git+https://github.com/sambit-giri/AstronomyCalc.git
 
-The dependencies should be installed automatically during the installation process. The list of required packages can be found in the requirements.txt file present in the root directory.
+*Important Notes*:
+    * Python Version: This package requires Python version 3.8 or higher.
+    * Virtual Environment: Using a Python virtual environment is recommended to prevent conflicts with system software. You can create one using Anaconda or venv.
+    * The dependencies should be installed automatically during the installation process. The list of required packages can be found in the requirements.txt file present in the root directory.
 
 ### Tests
 
