### 2024-12-24 b2bb5b5
notes
