### 2024-12-24 c7a3598
notes
