### 2024-12-24 b64af64
notebook usage
