### 2024-12-24 5e18f70
sentence fix
