+++ README.md
@@ -56,7 +56,7 @@ One can also install the latest version using pip by running the following comma
 
 ### Tests
 
-For testing, one can use [pytest](https://docs.pytest.org/en/stable/). To run all the test script, run the either of the following::
+For testing, one can use [pytest](https://docs.pytest.org/en/stable/). To run all the test script, run the following::
 
     python -m pytest tests
     
