### 2024-12-24 #261 @sambit-giri
@AstrobioMike Thank you so much!

>
>     * Noting what python version is required in the installation text would be useful. Prior to doing the pip install, often users will create a virtual environment (like you suggest in the readthedocs 👍), and knowing the required python base ahead of time can help save time. Adding a line like "If creating a virtual environment, it should use python X.X", or something.
>

I have added a list in the installation section of readme, which mentions the python version and a recommendation to use virtual environment.

>     * I don't know if it would cause problems with other dependencies, but if possible, you might consider adding jupyter to your package dependencies so after users install your package, they can immediately use the notebooks you provide with out needing to figure out installing something else. And/or just making a note somewhere how folks can setup an environment to use the notebooks could help. Though I also want to explicitly acknowledge that you've also already setup google colabs and mybinders, which is great!
>

This is great suggestion. I have added jupyter to the requirements file. I have added a section called `Jupyter Notebook Usage` with a detailed instruction.

>     * You might consider linking to the relevant notebooks on the readthedocs tutorial page. E.g., being able to click to download the model_universes.ipynb notebook from the https://astronomycalc.readthedocs.io/en/latest/examples/tutorials.html#Model-universes section would be convenient. Or just linking at the top, e.g., instead of just saying "This is a tutorial to show a usage of the AstronomyCalc package.", you could add to the end, "Corresponding notebooks can be found [here]" (linking-to-the-notebooks-directory).
>

I have added the link to the folder containing the notebooks to the top of the tutorial.

>     * Super-minor text thing: under the "Tests" section of the main README and in the readthedocs, the sentence says to "run _either_ of the following" but only lists one example. And you could also add a note that the user would need to be in the installed AstronomyCalc directory (since not everyone is familiar with how pytest works). (it would have been easier to suggest these text things as a pull request, but i didn't want to be listed as a contributor for just adjusting some wording, haha)
>

Nice catch! Initially I had nosetest also, but it did not make sense to have multiple options. I have fixed the sentence.
