### 2024-12-26 be7fbd2
sentence fix
