### 2025-01-20 #261 @arm61
Sorry for the delay @sambit-giri (lots of teaching recently). I was wondering about the "Usage Example"; it leads to a large code block in the paper that is maybe better served in the documentation. Thoughts?
