### 2025-01-21 #261 @sambit-giri
Hi @arm61, no problem at all—I completely understand! I agree that the "Usage Example" section contains a large code block. It is essentially a condensed version of the first tutorial already available in the documentation. If you agree, I can remove this section from the paper.
