+++ paper/paper.md
@@ -37,68 +37,6 @@ Designed with the variation theory of learning in mind, this package enables stu
 
 Furthermore, `AstronomyCalc` equips users with the tools needed to analyze real or simulated astronomical data, thereby providing a comprehensive learning experience that reflects the true nature of scientific inquiry in astronomy. This makes `AstronomyCalc` an invaluable resource for education in astronomy and cosmology, enhancing both the depth and quality of learning in these fields.
 
-# Usage Example
-
-To illustrate the functionality of `AstronomyCalc`, we will estimate cosmological distances in different model universes. Specifically, we will consider a dark energy-dominated universe (de Sitter model), a matter-dominated universe (Einstein-de Sitter model), and a universe based on the cosmology inferred from cosmic microwave background radiation measurements by the Planck satellite.
-
-```python
-import numpy as np 
-import matplotlib.pyplot as plt
-import AstronomyCalc as astro
-
-# Redshift bins
-zarr = np.logspace(-2, 1, 100)
-
-# Einsten-de Sitter universe
-cosmo_dict = {'Om': 1.0, 'Or': 0, 'Ok': 0, 'Ode': 0, 'h': 0.67}
-param_EdS = astro.param(cosmo=cosmo_dict)
-
-cdist_EdS = astro.comoving_distance(param_EdS, zarr)
-pdist_EdS = astro.proper_distance(param_EdS, zarr)
-
-# de Sitter universe
-cosmo_dict = {'Om': 0.0, 'Or': 0, 'Ok': 0, 'Ode': 1.0, 'h': 0.67}
-param_dS = astro.param(cosmo=cosmo_dict)
-
-cdist_dS = astro.comoving_distance(param_dS, zarr)
-pdist_dS = astro.proper_distance(param_dS, zarr)
-
-# Benchmark model or Planck universe
-cosmo_dict = {'Om': 0.31, 'Or': 0.0, 'Ok': 0.0, 'Ode': 0.69, 'h': 0.67}
-param = astro.param(cosmo=cosmo_dict)
-
-cdist = astro.comoving_distance(param, zarr)
-pdist = astro.proper_distance(param, zarr)
-
-# Plots
-fig, axs = plt.subplots(1,2,figsize=(10,4))
-axs[0].loglog(zarr, cdist_dS, ls='-.', label='de Sitter')
-axs[0].loglog(zarr, cdist, ls='-', label='Planck')
-axs[0].loglog(zarr, cdist_EdS, ls='--', label='Einstein-de Sitter')
-axs[0].legend(loc=2, fontsize=14)
-axs[0].set_xlabel('Redshift', fontsize=15)
-axs[0].set_ylabel('Comoving distance', fontsize=15)
-axs[0].axis([0.01,10,40,5e4])
-
-axs[1].loglog(zarr, pdist_dS, ls='-.', label='de Sitter')
-axs[1].loglog(zarr, pdist, ls='-', label='Planck')
-axs[1].loglog(zarr, pdist_EdS, ls='--', label='Einstein-de Sitter')
-axs[1].legend(loc=2, fontsize=14)
-axs[1].set_xlabel('Redshift', fontsize=15)
-axs[1].set_ylabel('Proper distance', fontsize=15)
-axs[1].axis([0.01,10,40,5e4])
-
-plt.tight_layout()
-plt.show()
-
-```
-
[…]
