+++ paper/paper.md
@@ -23,11 +23,11 @@ bibliography: paper.bib
 
 Understanding astrophysical and cosmological processes can be challenging due to their complexity and the lack of simple, everyday analogies. To address this, we present `AstronomyCalc`, a user-friendly Python package designed to facilitate the learning of these processes and help develop insights based on the variation theory of learning [@lo2011towards; @ling2012variation].
 
-`AstronomyCalc` enables students and educators to engage with key astrophysical and cosmological calculations, such as solving the Friedmann equations, which are fundamental to modeling the dynamics of the universe. The package allows users to construct and explore various cosmological models, including the de Sitter and Einstein-de Sitter universes [see @ryden2017introduction for more examples], by adjusting key parameters such as matter density and the Hubble constant. This interactive approach helps users intuitively grasp how variations in these parameters affect properties like expansion rates and cosmic time evolution.
+`AstronomyCalc` enables students and educators to engage with key astrophysical and cosmological calculations, such as solving the Friedmann equations, which are fundamental to modeling the dynamics of the universe. The package allows users to construct and explore various cosmological models, including the de Sitter and Einstein-de Sitter universes [see @ryden2017introduction for more examples], by adjusting key parameters such as matter density and the Hubble constant. This interactive approach helps users intuitively grasp how variations in these parameters affect properties like expansion rates and cosmic time evolution. This package written in such as manner that it can be easily expanded with more astronomical caluclations that is required in a course. The package is designed to be easily expanded with additional astronomical calculations as needed for a course.
 
 Moreover, `AstronomyCalc` includes modules for generating synthetic astronomical data or accessing publicly available datasets. In its current version, users can generate synthetic Type Ia supernova measurements of cosmological distances [@vanderplas2012introduction] or utilize the publicly available Pantheon+ dataset [@brout2022pantheonplus]. Additionally, the package supports the download and analysis of the SPARC dataset, which contains galaxy rotation curves for 175 disk galaxies [@lelli2016sparc].
 
-These datasets can be analyzed within the package to test cosmological and astrophysical models, offering a hands-on experience that mirrors the scientific research process in astronomy. `AstronomyCalc` implements simplified versions of advanced data analysis algorithms, such as the Metropolis-Hastings algorithm for Monte Carlo Markov chains [@robert2004metropolis], to explain the fundamental workings of these methods. By integrating theoretical concepts with observational data analysis, `AstronomyCalc` not only aids in conceptual learning but also provides insights into the empirical methods used in the field.
+The datasets provided in the package can be analyzed within the package to test cosmological and astrophysical models, offering a hands-on experience that mirrors the scientific research process in astronomy. `AstronomyCalc` implements simplified versions of advanced data analysis algorithms, such as the Importance sampling [@tokdar2010importance] and Metropolis-Hastings algorithm [@robert2004metropolis] for Monte Carlo Markov chain sampling or statistical data interpretation, to explain the fundamental workings of these methods. By integrating theoretical concepts with observational data analysis, `AstronomyCalc` not only aids in conceptual learning but also provides insights into the empirical methods used in the field.
 
 # Statement of Need
 
