+++ paper/paper.bib
@@ -71,6 +71,17 @@
   url="https://doi.org/10.1007/978-1-4757-4145-2_7"
 }
 
+@article{tokdar2010importance,
+  title={Importance sampling: a review},
+  author={Tokdar, Surya T and Kass, Robert E},
+  journal={Wiley Interdisciplinary Reviews: Computational Statistics},
+  volume={2},
+  number={1},
+  pages={54--60},
+  year={2010},
+  publisher={Wiley Online Library}
+}
+
 @phdthesis{giri2019tomographic,
   title={Tomographic studies of the 21-cm signal during reionization: Going beyond the power spectrum},
   author={Giri, Sambit K.},
