+++ paper/paper.bib
@@ -17,13 +17,6 @@
   publisher={G{\"o}teborg: Acta Universitatis Gothoburgensis}
 }
 
-@book{ryden2017introduction,
-  title={Introduction to cosmology},
-  author={Ryden, Barbara},
-  year={2017},
-  publisher={Cambridge University Press}
-}
-
 @inproceedings{vanderplas2012introduction,
   title={Introduction to astroML: Machine learning for astrophysics},
   author={VanderPlas, Jacob and Connolly, Andrew J and Ivezi{\'c}, {\v{Z}}eljko and Gray, Alex},
@@ -80,6 +73,13 @@
   pages={54--60},
   year={2010},
   publisher={Wiley Online Library}
+}
+
+@book{ryden2017introduction,
+  title={Introduction to cosmology},
+  author={Ryden, Barbara},
+  year={2017},
+  publisher={Cambridge University Press}
 }
 
 @phdthesis{giri2019tomographic,
