### 2025-01-22 46c3afd
a paragraph about notebooks
