### 2025-01-22 #261 @arm61
Fab, can you now mint a DOI using Zenodo for the manuscript and make a versioned release of the materials?
