### 2025-01-22 #261 @sambit-giri
I have created a new release on GitHub as AstronomyCalc v1.1 and submitted to Zenodo, producing DOI 10.5281/zenodo.14719952 for this package.
