### 2025-01-26 #261 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
