### 2025-02-12 c3ea4fc
unit fix in distance_modulus
