### 2025-02-12 a3f8335
colab link to bayesian_inference fixed
