+++ notebooks/README.md
@@ -6,5 +6,5 @@ Some of the examples have been set up to try online:
 
 * Model Universes: [Google Colab](https://colab.research.google.com/drive/1rLj1yaX8kJ_hIl8edGMJO75C6iiuk5si?usp=sharing), [MyBinder](https://mybinder.org/v2/gh/sambit-giri/AstronomyCalc/f7c00605caed709f38345d4f9101e4719468fda1?urlpath=lab%2Ftree%2Fnotebooks%2Fmodel_universes.ipynb)
 * Hubble Diagram: [Google Colab](https://colab.research.google.com/drive/1CRaVwn65-PhLah2iFjFSHCqCV37MI_1E?usp=sharing), [MyBinder](https://mybinder.org/v2/gh/sambit-giri/AstronomyCalc/6060e8e1875fa0d462478715feb80c8de7ff4dfb?urlpath=lab%2Ftree%2Fnotebooks%2Fhubble_diagram.ipynb)
-* Bayesian Parameter Inference: [Google Colab](https://drive.google.com/file/d/1QUkOuw9lRvlJ0crlbB9qhF70kbwRlTUb/view?usp=sharing), [MyBinder](https://mybinder.org/v2/gh/sambit-giri/AstronomyCalc/6060e8e1875fa0d462478715feb80c8de7ff4dfb?urlpath=lab%2Ftree%2Fnotebooks%2Fbayesian_inference.ipynb)
 * Rotation curves: [Google Colab](https://colab.research.google.com/drive/1EmAABcFnDG6EcEOY4ufs6wI14UT6LvWB?usp=sharing), [MyBinder](https://mybinder.org/v2/gh/sambit-giri/AstronomyCalc/6e5b63de3f71d4ee0f77b6c11956f6c5e0b28226?urlpath=lab%2Ftree%2Fnotebooks%2Frotation_curves.ipynb)
+* Bayesian Inference: [Google Colab](https://colab.research.google.com/drive/1TAXdiNLxJsC_LlAIzOYqQMwSueaJstj-?usp=sharing), [MyBinder](https://mybinder.org/v2/gh/sambit-giri/AstronomyCalc/6060e8e1875fa0d462478715feb80c8de7ff4dfb?urlpath=lab%2Ftree%2Fnotebooks%2Fbayesian_inference.ipynb)
