### 2025-02-12 a54fb9f
tests for data generation modules
