### 2025-02-12 da0bfab
test data removed
The following error only appears in CI GitHub Actions, but not when tested locally.
test_Hubble1929_data - AttributeError: module 'importlib.resources' has no attribute 'files'
