### 2025-03-03 6b1e9ae
backport importlib_resources for python<3.9
