+++ environment.yml
@@ -1,9 +1,9 @@
-name: toolscosmo
+name: astronomycalc
 channels:
   - defaults
   - conda-forge
 dependencies:
-  - python=3.8
+  - python=3.10
   - numpy
   - scipy
   - matplotlib
