### 2025-03-03 eb3f4b4
pkg_resources removed
