### 2025-03-03 865c644
CI with python=3.10
