+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.8]  # Test specifically with Python 3.8
+        python-version: [3.10]  # Test specifically with Python 3.10
 
     steps:
       - name: Check out repository code
