+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.10]  # Test specifically with Python 3.10
+        python-version: [3.9]  # Test specifically with Python 3.9
 
     steps:
       - name: Check out repository code
