### 2025-03-03 d7ceda9
again CI with python=3.10.*
