### 2025-05-04 #261 @labarba
The Zenodo archive displays a CC-BY license whereas the GitHub repository has an MIT license. Are the materials dual-licensed, given that they include code and non-code content? If so, maybe it should say so in the README.
