### 2025-05-04 #261 @labarba
You may want to use `{}` to caps-protect "Metropolis-Hastings" in a reference title; maybe "Pantheon+" too?
