### 2025-05-05 #261 @labarba
In Zenodo, you also can indicate dual-licensing.
