### 2025-05-07 a72b4af
Dual licensing clarified
