### 2025-05-07 3c4aefc
cap-protect of references
