+++ paper/paper.bib
@@ -28,7 +28,7 @@
 }
 
 @article{brout2022pantheonplus,
-  title={The Pantheon+ analysis: cosmological constraints},
+  title={The {Pantheon+} analysis: cosmological constraints},
   author={Brout, Dillon and Scolnic, Dan and Popovic, Brodie and Riess, Adam G and Carr, Anthony and Zuntz, Joe and Kessler, Rick and Davis, Tamara M and Hinton, Samuel and Jones, David and others},
   journal={The Astrophysical Journal},
   volume={938},
@@ -53,7 +53,7 @@
 
 @Inbook{robert2004metropolis,
   author="Robert, Christian P., and Casella, George",
-  title="The Metropolis---Hastings Algorithm",
+  title="The {Metropolis-Hastings} Algorithm",
   bookTitle="Monte Carlo Statistical Methods",
   year="2004",
   publisher="Springer New York",
