### 2025-05-07 #261 @sambit-giri
@labarba Thank you for the suggestions. I have updated the references in the paper and indicated the dual licensing in both the GitHub repository and Zenodo.
